requirements.txt
python-telegram-bot==21.11.1
requests==2.31.0
python-dotenv==1.0.0
Pillow==11.1.0
pytz==2025.1